Add `engine/collection.py` for the `Collection` newtype
[ci skip-rust-tests]
[ci skip-jvm-tests]
