+++ src/python/pants/backend/project_info/rules/BUILD
@@ -5,8 +5,8 @@ python_library(
   dependencies=[
     '3rdparty/python:dataclasses',
     'src/python/pants/engine:console',
+    'src/python/pants/engine:collection',
     'src/python/pants/engine:fs',
-    'src/python/pants/engine:objects',
     'src/python/pants/engine:rules',
     'src/python/pants/engine:selectors',
     'src/python/pants/engine/legacy:graph',
