+++ src/python/pants/backend/python/rules/BUILD
@@ -15,6 +15,7 @@ python_library(
     'src/python/pants/engine/legacy:graph',
     'src/python/pants/engine:addresses',
     'src/python/pants/engine:build_files',
+    'src/python/pants/engine:collection',
     'src/python/pants/engine:fs',
     'src/python/pants/engine:isolated_process',
     'src/python/pants/engine:rules',
