+++ src/python/pants/base/BUILD
@@ -148,8 +148,8 @@ python_library(
   sources = ['specs.py'],
   dependencies = [
     '3rdparty/python:dataclasses',
+    'src/python/pants/engine:collection',
     'src/python/pants/engine:fs',
-    'src/python/pants/engine:objects',
     'src/python/pants/util:collections',
     'src/python/pants/util:dirutil',
     'src/python/pants/util:filtering',
