+++ src/python/pants/core/goals/BUILD
@@ -9,13 +9,13 @@ python_library(
     'src/python/pants/base:specs',
     'src/python/pants/core/util_rules',
     'src/python/pants/engine:addresses',
+    'src/python/pants/engine:collection',
     'src/python/pants/engine:console',
     'src/python/pants/engine:desktop',
     'src/python/pants/engine:fs',
     'src/python/pants/engine:goal',
     'src/python/pants/engine:interactive_runner',
     'src/python/pants/engine:isolated_process',
-    'src/python/pants/engine:objects',
     'src/python/pants/engine:rules',
     'src/python/pants/engine:selectors',
     'src/python/pants/engine:target',
