+++ src/python/pants/core/util_rules/BUILD
@@ -9,8 +9,8 @@ python_library(
     'src/python/pants/base:specs',
     'src/python/pants/core',
     'src/python/pants/engine:addresses',
+    'src/python/pants/engine:collection',
     'src/python/pants/engine:fs',
-    'src/python/pants/engine:objects',
     'src/python/pants/engine:rules',
     'src/python/pants/engine:selectors',
     'src/python/pants/engine:target',
