+++ src/python/pants/engine/collection_test.py
@@ -1,7 +1,7 @@
 # Copyright 2019 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.engine.objects import Collection
+from pants.engine.collection import Collection
 
 
 class Examples(Collection[int]):
