+++ src/python/pants/engine/legacy/graph.py
@@ -30,6 +30,7 @@ from pants.build_graph.build_graph import BuildGraph
 from pants.build_graph.remote_sources import RemoteSources
 from pants.build_graph.target import Target as TargetV1
 from pants.engine.addresses import Addresses, AddressesWithOrigins, AddressWithOrigin
+from pants.engine.collection import Collection
 from pants.engine.fs import EMPTY_SNAPSHOT, PathGlobs, Snapshot
 from pants.engine.legacy.address_mapper import LegacyAddressMapper
 from pants.engine.legacy.structs import (
@@ -39,7 +40,6 @@ from pants.engine.legacy.structs import (
     SourcesField,
     TargetAdaptor,
 )
-from pants.engine.objects import Collection
 from pants.engine.parser import HydratedStruct
 from pants.engine.rules import RootRule, rule
 from pants.engine.selectors import Get, MultiGet
