+++ tests/python/pants_test/engine/test_mapper.py
@@ -11,9 +11,9 @@ from pants.base.specs import AddressSpec, AddressSpecs, SingleAddress
 from pants.build_graph.address import Address
 from pants.engine.addresses import Addresses
 from pants.engine.build_files import create_graph_rules
+from pants.engine.collection import Collection
 from pants.engine.fs import create_fs_rules
 from pants.engine.mapper import AddressFamily, AddressMap, AddressMapper, DifferingFamiliesError
-from pants.engine.objects import Collection
 from pants.engine.parser import HydratedStruct, SymbolTable
 from pants.engine.rules import rule
 from pants.engine.selectors import Get, MultiGet
